jackson2: deprecate JacksonModule in favor of Jackson2Module
- fix #3898
