+++ docs/asciidoc/modules/tRPC.adoc
@@ -17,7 +17,7 @@ Installing tRPC in your Jooby app is a 3-step process. Because tRPC relies heavi
 [source,java,role="primary"]
 ----
 import io.jooby.Jooby;
-import io.jooby.json.JacksonModule;
+import io.jooby.json.Jackson2Module;
 import io.jooby.trpc.TrpcModule;
 import io.jooby.trpc.TrpcJackson2Module;
 
@@ -34,7 +34,7 @@ import io.jooby.trpc.TrpcJackson2Module;
 [source,kotlin,role="secondary"]
 ----
 import io.jooby.kt.Kooby
-import io.jooby.json.JacksonModule
+import io.jooby.json.Jackson2Module
 import io.jooby.trpc.TrpcModule
 import io.jooby.trpc.TrpcJackson2Module
 
@@ -47,7 +47,7 @@ import io.jooby.trpc.TrpcJackson2Module
 }
 ----
 
-1. Install a supported JSON engine (`Jackson3Module`, `JacksonModule` for Jackson 2, or `AvajeJsonbModule`).
+1. Install a supported JSON engine (`Jackson3Module`, `Jackson2Module` for Jackson 2, or `AvajeJsonbModule`).
 2. Install the corresponding bridge module (`TrcJackson3Module`, `TrpcJackson2Module`, or `TrpcAvajeJsonbModule`).
 3. Install the core tRPC extension (`TrpcModule`).
 4. Register your `@Trpc` annotated controllers (using the APT generated route).
