+++ modules/jooby-jackson/src/main/java/module-info.java
@@ -3,7 +3,52 @@
  * Apache License Version 2.0 https://jooby.io/LICENSE.txt
  * Copyright 2014 Edgar Espina
  */
-/** Jackson module. */
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+/**
+ * JSON module using Jackson: https://jooby.io/modules/jackson2.
+ *
+ * <p>Usage:
+ *
+ * <pre>{@code
+ * {
+ *
+ *   install(new Jackson2Module());
+ *
+ *   get("/", ctx -> {
+ *     MyObject myObject = ...;
+ *     // send json
+ *     return myObject;
+ *   });
+ *
+ *   post("/", ctx -> {
+ *     // read json
+ *     MyObject myObject = ctx.body(MyObject.class);
+ *     // send json
+ *     return myObject;
+ *   });
+ * }
+ * }</pre>
+ *
+ * For body decoding the client must specify the <code>Content-Type</code> header set to <code>
+ * application/json</code>.
+ *
+ * <p>You can retrieve the {@link ObjectMapper} via require call:
+ *
+ * <pre>{@code
+ * {
+ *
+ *   ObjectMapper mapper = require(ObjectMapper.class);
+ *
+ * }
+ * }</pre>
+ *
+ * Complete documentation is available at: https://jooby.io/modules/jackson2.
+ *
+ * @author edgar
+ * @since 2.0.0
+ */
 module io.jooby.jackson {
   exports io.jooby.jackson;
