+++ tests/src/test/java/io/jooby/i2818/Issue2818.java
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
 
-import io.jooby.jackson.JacksonModule;
+import io.jooby.jackson.Jackson2Module;
 import io.jooby.junit.ServerTest;
 import io.jooby.junit.ServerTestRunner;
 
@@ -117,7 +117,7 @@ public class Issue2818 {
     runner
         .define(
             app -> {
-              app.install(new JacksonModule());
+              app.install(new Jackson2Module());
               app.ws(
                   "/ws/render-bin",
                   (ctx, initializer) -> {
@@ -139,7 +139,7 @@ public class Issue2818 {
     runner
         .define(
             app -> {
-              app.install(new JacksonModule());
+              app.install(new Jackson2Module());
               app.ws(
                   "/ws/render-bin",
                   (ctx, initializer) -> {
