+++ tests/src/test/java/io/jooby/i2923/Issue2923.java
@@ -7,7 +7,7 @@ package io.jooby.i2923;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import io.jooby.jackson.JacksonModule;
+import io.jooby.jackson.Jackson2Module;
 import io.jooby.junit.ServerTest;
 import io.jooby.junit.ServerTestRunner;
 
@@ -18,7 +18,7 @@ public class Issue2923 {
     runner
         .define(
             app -> {
-              app.install(new JacksonModule());
+              app.install(new Jackson2Module());
               app.get(
                   "/2923",
                   ctx -> {
