+++ tests/src/test/java/io/jooby/i3508/Issue3508.java
@@ -24,7 +24,7 @@ import io.jooby.i3508.data.AvajeNewAccountRequest;
 import io.jooby.i3508.data.HbvNewAccountRequest;
 import io.jooby.i3508.data.NewAccountRequest;
 import io.jooby.i3508.data.Person;
-import io.jooby.jackson.JacksonModule;
+import io.jooby.jackson.Jackson2Module;
 import io.jooby.junit.ServerTest;
 import io.jooby.junit.ServerTestRunner;
 import io.jooby.validation.BeanValidator;
@@ -41,7 +41,7 @@ public class Issue3508 {
     runner
         .define(
             app -> {
-              app.install(new JacksonModule());
+              app.install(new Jackson2Module());
               app.install(new HibernateValidatorModule());
 
               app.get("/3508/query", validate(ctx -> ctx.query().to(Bean3508.class)));
