+++ tests/src/test/java/io/jooby/i3868/Jackson2JsonRpcProtocolTest.java
@@ -6,13 +6,13 @@
 package io.jooby.i3868;
 
 import io.jooby.Jooby;
-import io.jooby.jackson.JacksonModule;
+import io.jooby.jackson.Jackson2Module;
 import io.jooby.jsonrpc.jackson2.JsonRpcJackson2Module;
 
 public class Jackson2JsonRpcProtocolTest extends AbstractJsonRpcProtocolTest {
   @Override
   protected void installJsonEngine(Jooby app) {
-    app.install(new JacksonModule());
+    app.install(new Jackson2Module());
     app.install(new JsonRpcJackson2Module());
   }
 }
