+++ tests/src/test/kotlin/i3476/Issue3476.kt
@@ -5,7 +5,7 @@
  */
 package i3476
 
-import io.jooby.jackson.JacksonModule
+import io.jooby.jackson.Jackson2Module
 import io.jooby.junit.ServerTest
 import io.jooby.junit.ServerTestRunner
 import io.jooby.kt.Kooby
@@ -17,7 +17,7 @@ class Issue3476 {
     runner
       .use {
         Kooby {
-          install(JacksonModule())
+          install(Jackson2Module())
           mvc(C3476_())
         }
       }
